기능구현 오류수정및 추가

+++ backend/src/main/java/com/backend/service/store/PaymentService.java
@@ -25,22 +25,33 @@ public class PaymentService {
 
         mapper.add(payment);
 
-        for (int i = 0; i < payment.getCheckCartId().size(); i++) {
+        if (payment.getCheckCartId() != null && !payment.getCheckCartId().isEmpty()) {
 
-            orderMapper.copyCartData(payment.getCheckCartId().get(i), payment.getId());
+            for (int i = 0; i < payment.getCheckCartId().size(); i++) {
+
+                orderMapper.copyCartData(payment.getCheckCartId().get(i), payment.getId());
+            }
+
+            for (int i = 0; i < payment.getCheckCartId().size(); i++) {
+
+                Integer getQuantity = cartMapper.getQuantity(payment.getCheckCartId().get(i));
+                Integer productId = cartMapper.getProductId(payment.getCheckCartId().get(i));
+                productMapper.updateStock(productId, getQuantity);
+            }
+
+            for (int i = 0; i < payment.getCheckCartId().size(); i++) {
+
+                cartMapper.deleteCartByCheckCartId(payment.getCheckCartId().get(i));
+            }
+        } else {
+
+            System.out.println("payment.getBuyerDate() = " + payment.getBuyerDate());
+
+            orderMapper.addSinggleProductOrder(payment.getProductId(), payment.getQuantity(), payment.getId(), payment.getName(), payment.getAmount(), payment.getAmount(), payment.getMemberNumber());
+
+            productMapper.updateStock(payment.getProductId(), payment.getQuantity());
         }
 
-        for (int i = 0; i < payment.getCheckCartId().size(); i++) {
-
-            Integer getQuantity = cartMapper.getQuantity(payment.getCheckCartId().get(i));
-            Integer productId = cartMapper.getProductId(payment.getCheckCartId().get(i));
-            productMapper.updateStock(productId, getQuantity);
-        }
-
-        for (int i = 0; i < payment.getCheckCartId().size(); i++) {
-
-            cartMapper.deleteCartByCheckCartId(payment.getCheckCartId().get(i));
-        }
 
         return payment.getId();
     }
